refactor: #12 MyPageService 변경 및 정리
