+++ src/main/java/com/example/creditmarket/service/FavorService.java
@@ -1,0 +1,11 @@
+package com.example.creditmarket.service;
+
+import com.example.creditmarket.dto.response.FavoriteListResponseDTO;
+
+public interface FavorService {
+
+    String favoriteService(String productId, String userEmail);
+
+    FavoriteListResponseDTO selectFavoriteList(int page, String userEmail);
+
+}
