--- src/main/java/com/example/creditmarket/service/MyPageService.java
@@ -1,14 +1,0 @@
-package com.example.creditmarket.service;
-
-import com.example.creditmarket.dto.response.FavoriteListResponseDTO;
-import com.example.creditmarket.dto.response.OrderListResponseDTO;
-
-public interface MyPageService {
-
-    FavoriteListResponseDTO selectFavoriteList(int page, String userEmail);
-
-    OrderListResponseDTO selectOrderList(int page, String userEmail);
-
-    String updateOrder(Long orderId, String userEmail);
-
-}
